http-netty: validate the HTTP/2 ':path' pseudo-header
Motivation:
Nothing validated ':path' characters, so a peer could send CR, LF, NUL or
SP in it and we would hand it to the service and re-emit it on an
HTTP/1.1 downgrade - the header-injection route RFC 9113, 8.2.1 calls out
as request smuggling. Outbound we sent an empty ':path' where 8.3.1
requires "/", and HPACK silently mangled any non-ASCII target.

Modifications:
- Validate ':path' in both duplex handlers: non-empty, VCHAR only
(0x21-0x7e), per RFC 9113, 8.3.1.
- Inbound violations become a PROTOCOL_ERROR stream error; outbound ones
fail the write before the value reaches HPACK.
- Send "/" for an empty request-target, as HTTP/1.x already does.

Result:
An illegal ':path' no longer crosses the HTTP/2 boundary in either
direction, and an empty request-target is sent as "/" on both protocols.

A raw non-ASCII target now fails instead of being silently corrupted; use
path(String) or requestTarget(String, Charset) to percent-encode, which
the request-target API already requires. HTTP/1.x is unchanged.

+++ servicetalk-http-netty/src/main/java/io/servicetalk/http/netty/HttpObjectDecoder.java
@@ -1081,7 +1081,7 @@ abstract class HttpObjectDecoder<T extends HttpMetaData> extends ByteToMessageDe
         return value == SP || value == HT;
     }
 
-    static boolean isVCHAR(final byte value) {
+    static boolean isVCHAR(final int value) {
         return value >= '!' && value <= '~';
     }
